pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "library"
version = "0.1.0"
description = "A FastAPI-based library management system"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.4.2",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.23",
    "alembic>=1.12.1",
    "psycopg2-binary>=2.9.9",
    "httpx>=0.26.0",
    "asyncpg>=0.29.0"
]
requires-python = ">=3.9"
readme = "README.md"
license = { text = "MIT" }

[project.urls]
Homepage = "https://github.com/math-s/library"
Documentation = "https://github.com/math-s/library#readme"
Repository = "https://github.com/math-s/library.git"

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "sqlalchemy-utils>=0.41.1",
    "docker>=7.0.0",
    "pre-commit>=3.6.0",
    "black>=24.2.0"
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=app --cov-report=term-missing"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["app"]
omit = ["tests/*", "app/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError"
]

[tool.ruff]
line-length = 88
target-version = "py39"
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]
ignore = []

[tool.ruff.isort]
known-first-party = ["app"]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$' 